Add config support
