--- crates/config/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "janitor-config"
-version = "0.0.0"
-authors = ["Jelmer Vernooĳ <jelmer@jelmer.uk>"]
-edition = "2021"
-description = "Config for the janitor"
-license = "GPL-3.0+"
-repository = "https://github.com/jelmer/janitor.git"
-homepage = "https://github.com/jelmer/janitor"
-build = "build.rs"
-
-[dependencies]
-protobuf = "3"
-
-[build-dependencies]
-protobuf-codegen = "3"
-protoc-rust = "^2.0"
