--- crates/config/src/lib.rs
@@ -1,29 +1,0 @@
-include!(concat!(env!("OUT_DIR"), "/generated/mod.rs"));
-
-use protobuf::text_format;
-use std::fs::File;
-use std::io::Read;
-
-pub use config::{AptRepository, Campaign, Config, Distribution};
-
-pub fn read_config(file_path: &str) -> Result<Config, Box<dyn std::error::Error>> {
-    let mut file = File::open(file_path)?;
-    let mut contents = String::new();
-    file.read_to_string(&mut contents)?;
-
-    Ok(text_format::parse_from_str(contents.as_str())?)
-}
-
-impl Config {
-    pub fn get_distribution(&self, name: &str) -> Option<&Distribution> {
-        self.distribution
-            .iter()
-            .find(|d| d.name.as_ref().unwrap() == name)
-    }
-
-    pub fn get_campaign(&self, name: &str) -> Option<&Campaign> {
-        self.campaign
-            .iter()
-            .find(|c| c.name.as_ref().unwrap() == name)
-    }
-}
